Android implementation cleanup
