+++ android/src/main/java/com/swmansion/gesturehandler/react/eventbuilders/HoverGestureHandlerEventDataBuilder.kt
@@ -2,19 +2,23 @@ package com.swmansion.gesturehandler.react.eventbuilders
 
 import com.facebook.react.bridge.WritableMap
 import com.facebook.react.uimanager.PixelUtil
+import com.swmansion.gesturehandler.core.GestureUtils
 import com.swmansion.gesturehandler.core.HoverGestureHandler
+import com.swmansion.gesturehandler.core.StylusData
 
 class HoverGestureHandlerEventDataBuilder(handler: HoverGestureHandler) : GestureHandlerEventDataBuilder<HoverGestureHandler>(handler) {
   private val x: Float
   private val y: Float
   private val absoluteX: Float
   private val absoluteY: Float
+  private val stylusData: StylusData
 
   init {
     x = handler.lastRelativePositionX
     y = handler.lastRelativePositionY
     absoluteX = handler.lastPositionInWindowX
     absoluteY = handler.lastPositionInWindowY
+    stylusData = handler.stylusData
   }
 
   override fun buildEventData(eventData: WritableMap) {
@@ -25,6 +29,7 @@ class HoverGestureHandlerEventDataBuilder(handler: HoverGestureHandler) : Gestur
       putDouble("y", PixelUtil.toDIPFromPixel(y).toDouble())
       putDouble("absoluteX", PixelUtil.toDIPFromPixel(absoluteX).toDouble())
       putDouble("absoluteY", PixelUtil.toDIPFromPixel(absoluteY).toDouble())
+      putMap("stylusData", GestureUtils.createStylusDataObject(stylusData))
     }
   }
 }
